Update chargelnd-htlc.config
zap-o-matic is no longer updating max htlc every 3 hours--was too gossip heavy and with lots of transactions, it was just getting out of date, so now max-htlc is set to 95% of the channel capacity (where channel policy allows) and channels disable when they get to 3%. Throwing this out there in case anyone else is auto-updating HTLCs with charge-lnd on a cron-job. Probably more harm than good once you start routing a ton.